Move all function definitions out to separate files.
