+++ src/classes/atomType.cpp
@@ -6,34 +6,6 @@
 #include "templates/algorithms.h"
 #include <map>
 
-// Return enum options for ShortRangeType
-EnumOptions<ShortRangeFunctions::Form> ShortRangeFunctions::forms()
-{
-    return EnumOptions<ShortRangeFunctions::Form>("ShortRangeType",
-                                                  {{ShortRangeFunctions::Form::None, "None"},
-                                                   {ShortRangeFunctions::Form::LennardJones, "LJ", 2, 2},
-                                                   {ShortRangeFunctions::Form::LennardJonesGeometric, "LJGeometric", 2, 2}});
-}
-
-// Return parameters for specified form
-const std::vector<std::string> &ShortRangeFunctions::parameters(Form form)
-{
-    static std::map<ShortRangeFunctions::Form, std::vector<std::string>> params_ = {
-        {ShortRangeFunctions::Form::None, {}},
-        {ShortRangeFunctions::Form::LennardJones, {"epsilon", "sigma"}},
-        {ShortRangeFunctions::Form::LennardJonesGeometric, {"epsilon", "sigma"}}};
-    return params_[form];
-}
-
-// Return nth parameter for the given form
-std::string ShortRangeFunctions::parameter(Form form, int n) { return DissolveSys::stringAt(parameters(form), n); }
-
-// Return index of parameter in the given form
-std::optional<int> ShortRangeFunctions::parameterIndex(Form form, std::string_view name)
-{
-    return DissolveSys::indexOf(parameters(form), name);
-}
-
 AtomType::AtomType(Elements::Element Z) : Z_(Z), interactionPotential_(ShortRangeFunctions::Form::None) {}
 AtomType::AtomType(std::string_view name) : name_(name), interactionPotential_(ShortRangeFunctions::Form::None) {}
 AtomType::AtomType(Elements::Element Z, std::string_view name)
